+++ storybook/.storybook/main.js
@@ -2,12 +2,7 @@ const TsconfigPathsPlugin = require('tsconfig-paths-webpack-plugin')
 
 module.exports = {
   stories: ['../src/**/*.stories.@(ts|tsx|js|jsx|mdx)'],
-  addons: [
-    '@storybook/addon-a11y/register',
-    '@storybook/addon-storysource',
-    '@storybook/addon-backgrounds/register',
-    '@storybook/addon-knobs/register',
-  ],
+  addons: ['@storybook/addon-essentials'],
 
   webpackFinal: async (config) => {
     config.module.rules.push({
@@ -16,10 +11,6 @@ module.exports = {
         {
           loader: require.resolve('babel-loader'),
         },
-        // // Optional
-        // {
-        //   loader: require.resolve('react-docgen-typescript-loader'),
-        // },
       ],
     })
     config.resolve.extensions.push('.ts', '.tsx')